Add files via upload
This file I added the css style sheet  with our style stuff and linked it to the html document. I did this because the assignment said : A css stylesheet linked to from that index".
I also added in an image of the wireframes of our basic website idea just in the folder. I think the assignment doc said that dr dan wanted that too.